SSUTO-86 TASK: Deprecate Eureka Service Discovery
This change commits to using service discovery created by Docker Compose
for ECS deployments. For local development, no service discovery will be
necessary. Instead, we are providing an additional `local` profile for
interacting with the configuration server. As such, the Eureka dependency
has been removed, and the `ecs` profile Bean for handling Eureka has been
removed.

As this service uses a FeignClient for interacting with the AuthService,
an additional client is provided as an extension of the ECS FeignClient
for local development.

This additionally excludes SLF4J bindings from the SpotBugs dependency
which was causing a warning on application boot.